環境によってモチベーション値が小数点以下第一位まで表示されてしまうのを修正

+++ scheduler/lib/model/schedule.dart
@@ -1,6 +1,6 @@
 class Schedule {
   String name = "";
-  double motivation = 0.0;
+  int motivation = 0;
 
   DateTime startDateTime = DateTime(DateTime.now().year, DateTime.now().month,
       DateTime.now().day, DateTime.now().hour, DateTime.now().minute);
